# BUG: fix a regression in astropy 6.0.1 where using a dimensionless Quantity as an exponent could raise an exception