@neutrinoceros 2024-04-04
this was triaged to 6.0.2 bug the backport label wasn't added, so I'll assume it was a mistake and I'll open a manual backport PR