@pllim 2024-04-04
FWIW

```
| Change   | Before [495efdb3]    | After [7df147cb]    |   Ratio | Benchmark (Parameter)                                      |
|----------|----------------------|---------------------|---------|------------------------------------------------------------|
| -        | 300±6μs              | 227±4μs             |    0.76 | units.TimeQuantityOpLargeArraySameUnit.time_quantity_equal |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```